requirements/common.txt
aiosmtplib==1.1.4
asgiref==3.2.10
astroid==2.4.2
asyncio==3.4.3
celery==4.4.7
cffi==1.14.0
cryptography==2.9.2
dj-database-url==0.5.0
Django==3.0.8
django-cors-headers==3.5.0
django-filter==2.3.0
django-import-export==2.3.0
django-rest-knox==4.1.0
djangorestframework==3.11.0
gunicorn==20.0.4
isort==5.1.1
lazy-object-proxy==1.5.0
Markdown==3.2.2
mccabe==0.6.1
psycopg2-binary==2.8.5
pycparser==2.20
pylint==2.5.3
pytz==2020.1
requests==2.24.0
six==1.15.0
smtp==0.1.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
waitress==1.4.4
whitenoise==5.1.0
wrapt==1.12.1
xmltodict==0.12.0